startMoving and StopMoving methods added to AnimationManager.java

+++ src/main/java/entity/AnimationState.java
@@ -3,7 +3,7 @@ package entity;
 import com.project.engine.Rendering.SpriteRenderer;
 
 public interface AnimationState {
-    void onStart(SpriteRenderer renderer, int x, int y);   // Pass position
-    void onContinue(SpriteRenderer renderer, int x, int y);
-    void onEnd(SpriteRenderer renderer, int x, int y);
+    void onStart(SpriteRenderer renderer, double x, double y);   // Pass position
+    void onContinue(SpriteRenderer renderer, double x, double y);
+    void onEnd(SpriteRenderer renderer, double x, double y);
 }
